Reduce memory usage in ThumbnailCreator
When there are a lot of images to thumbnail, it's best not to load them
all into RAM first. In fact, we don't really need to load them at all in
order to thumbnail, just shell out to `convert`. That makes the ruby
garbage collector's job a bit easier.

In my local tests this keeps kontrast's memory usage around 500MB
instead of spiking up to 2.5GB.

Refactor ThumbnailCreator as module functions instead of mixin and add a
spec.